Spell, UI Refactor

+++ src/lib/utils/operations.ts
@@ -1,5 +1,6 @@
 import { supabaseClient } from "$lib/db";
 import Toastify from 'toastify-js';
+import type { ItemType, Spell } from "../../../types/types";
 
 export const saveEntity = async (entityType: string, entity) => {
   const entityToSave = { ...entity };
@@ -12,6 +13,11 @@ export const saveEntity = async (entityType: string, entity) => {
   }
   if (typeof entityToSave.specialties === 'string') {
     entityToSave.specialties = entityToSave.specialties.split(',').map((t) => t.trim());
+  }
+
+  // if entityType is spells, rename to spells_v2
+  if (entityType === 'spells') {
+    entityType = 'spells_v2';
   }
 
   // remove spells and equipment
@@ -54,6 +60,10 @@ export const copyEntity = async (entityType, entity) => {
   delete entityData.spells;
   delete entityData.equipment;
 
+  if (entityType === 'spells') {
+    entityType = 'spells_v2';
+  }
+
   const { data, error } = await supabaseClient.from(entityType).insert([entityData]).select();
 
   let location = "/"
@@ -76,12 +86,26 @@ export const copyEntity = async (entityType, entity) => {
   window.location.href = location
 }
 
-export const deleteEntity = async (entityType, entity) => {
+export const deleteEntity = async (entityType: ItemType, entity: Spell | any) => {
+
+  // alert window to confirm
+  const confirmed = confirm(`Are you sure you want to delete ${entity.name}?`);
+  if (!confirmed) {
+    return;
+  }
+
+  if (entityType === 'spells') {
+    entityType = 'spells_v2';
+  }
+
   const { error } = await supabaseClient.from(entityType).delete().eq('id', entity.id);
 
-  let location = "/"
+  let location = `/${entityType}`
   if (entityType == 'monsters') {
     location = "/npcs"
+  }
+  if (entityType == 'spells_v2') {
+    location = "/spells"
   }
 
   if (!error) {
@@ -105,11 +129,16 @@ export const createEntity = async (entityType, entity) => {
   delete entityData.id;
   entityData.user_id = userData?.user?.id;
 
+  if (entityType === 'spells') {
+    entityType = 'spells_v2';
+  }
+
+
   const { data, error } = await supabaseClient.from(entityType).insert([entityData]).select();
 
   let location = "/"
   if (entityType === 'monsters') {
-    location = `/npcs/${data[0].id}`
+    location = `/ npcs / ${data[0].id}`
   }
 
   if (!error) {
